requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alembic==1.3.3            # via flask-migrate
appdirs==1.4.3            # via black
attrs==19.3.0             # via black, pytest
beautifulsoup4==4.8.2     # via webtest
black==19.10b0
blinker==1.4              # via flask-debugtoolbar
cffi==1.13.2              # via cryptography
click==7.0
coverage==5.0.3           # via pytest-cov
cryptography==2.8
entrypoints==0.3          # via flake8
environs==7.1.0
factory-boy==2.12.0
faker==4.0.0
flake8==3.7.9
flask-debugtoolbar==0.10.1
flask-marshmallow==0.10.1
flask-migrate==2.5.2
flask-sqlalchemy==2.4.1
flask==1.1.1
gevent==1.4.0
greenlet==0.4.15          # via gevent
gunicorn==20.0.4
importlib-metadata==1.5.0  # via pluggy, pytest
itsdangerous==1.1.0       # via flask, flask-debugtoolbar
jinja2==2.11.1            # via flask
mako==1.1.1               # via alembic
markupsafe==1.1.1         # via jinja2, mako
marshmallow-sqlalchemy==0.21.0
marshmallow==3.3.0        # via environs, flask-marshmallow, marshmallow-sqlalchemy
mccabe==0.6.1             # via flake8
more-itertools==8.2.0     # via pytest
packaging==20.1           # via pytest
pathspec==0.7.0           # via black
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.4
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pycparser==2.19           # via cffi
pyflakes==2.1.1           # via flake8
pyjwt==1.7.1
pyparsing==2.4.6          # via packaging
pytest-cov==2.8.1
pytest==5.3.5
python-dateutil==2.8.1    # via alembic, faker
python-dotenv==0.10.5     # via environs
python-editor==1.0.4      # via alembic
regex==2020.1.8           # via black
six==1.14.0               # via cryptography, flask-marshmallow, packaging, python-dateutil, webtest
soupsieve==1.9.5          # via beautifulsoup4
sqlalchemy==1.3.13
text-unidecode==1.3       # via faker
toml==0.10.0              # via black
typed-ast==1.4.1          # via black
waitress==1.4.2           # via webtest
wcwidth==0.1.8            # via pytest
webob==1.8.6              # via webtest
webtest==2.0.34
werkzeug==0.16.1
zipp==2.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
